Refactor to class-based approach (1/2)
After some discussion, we decided to decouple three concepts so
we can respond to changing business concerns. Instead of a single
function that runs the simulation, we now have three classes:

> DeliveryPerson

A deliveryPerson starts at one location, and can act on a set of directions,
keeping track of the locations visited by the given directions.

> TakeTurnsDispatcher

A Dispatcher takes a list of DeliveryPerson on construction, and can
dispatch a set of directions to said deliveryPeople. The TakeTurnsDispatcher
uses a simple strategy of splitting up the given directions, turn by turn,
to each deliveryPerson.

> LocationsVisitedReducer

A Reducer takes an array of DeliveryPerson, and computes some statistic on it.
The LocationsVisitedReducer takes the history of locations the deliveryPeople have
visited and returns the number of unique houses that have been visited.

*Note* In this commit we have kept the `deliver` function as an export that
emits a warning if used. This will give a chance for callers to migrate to
the new code on next push.